safari-600.7-branch/LayoutTests/inspector-obsolete/styles/styles-add-invalid-property-expected.txt

Tests that adding an invalid property retains its syntax.

Text
Before append:
[expanded] 
element.style  { ()
font-size: 12px;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;


After append:
[expanded] 
element.style  { ()
font-size: 12px;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;
